Make improvements to Licensing
The license to the code is the same, but the following changes are made:

* My code is now more separated from code I was borrowing from Hex
Fiend which retains its copyright header

* I removed “Created by” comment on all source files and replaced
zgcoder with my own name. Should make it easier to use the source files
onto another project.

* I adjusted the LICENSE file to include licenses of all works that are
being used, along with mentioning the copyright on artwork assets.
(Localizations are still todo).

* I added an improved About Window where the user can read all credits
& licensing details.

* I removed one old stray GPL file that was not being included in the
Xcode project or used anywhere else, not being linked in the built code
at all.