Infer and allow changing disassembler mode
For Rosetta processes, the user may be looking at x86_64 or arm64 disassembly. We now by default infer if the code that is being examined is x86_64 or arm64 (by seeing if the calculated memory address is within __TEXT segment) and allow the user to change disassembler mode manually from the Debug menu from the "Interpreted Architecture" submenu.