Add code injection for arm64 using relative branching
This does not yet support far branches. Unfortunately far branches are > 128 MiB away for arm64 as opposed to INT32_MAX for x86. This means this relative branching support for code injection will only work for the most trivial kinds of applications (e.g. a simple command line tool).

We will need to think about another approach for far away branches for code injection to even be practically useful on this architecture.