Disable partial stepping for emulated breakpoints
We currently can't step over or out to a target emulated breakpoint, although we can step into a targeted emulated breakpoint. In the latter case, we then cannot step in or over the current emulated breakpoint.

The partial disabling logic is handled in the debugger controller UI rather than the breakpoint controller. Maybe some day we can support stepping with emulated breakpoints better.